src/aifrontend/requirements.txt
Flask==3.0.3
requests==2.32.3
gunicorn==22.0.0
PyJWT==2.9.0
cryptography==43.0.1
